Removed constructor, modified `setupAudioSystem()` and removed pause functionality (accidentally). I removed the custom constructor, but may need to return it; I removed it to test if not loading an audio file on startup saved some time in loading the GUI; alas, this was not so.
I modified `setupAudioSystem()` to decode the input file first; if the
file is already a WAV file, the decoding step is superfluous, so I shoud
add some way of detecting and accounting for that.

Pause functionality was removed as a consequence of removing the
constructor; I had called setupAudioSystem from within the constructor,
so I naievly moved it to `playFile()`. This means that pausing and
stopping are now synonomous.